autenticacion de rol

+++ src/controllers/CourseController.ts
@@ -1,34 +1,42 @@
 import { Request, Response } from 'express';
 import { Course } from '../models/Course';
+import { CustomError } from '../models/custom-error.model';
+import { rol } from '../models/Rol';
+
 import { CourseService } from '../services/CourseService';
 
 export class CourseController {
-  static async getCourses(_req: Request, res: Response) {
-    try {
-      const data = await CourseService.getCourses();
-      res.status(200).json(data);
-    } catch (error: any) {
-      throw new Error(error.message);
-    }
-  }
-
   static async getCourseById(req: Request, res: Response) {
     try {
       const idCourse = Number(req.params.id);
       const data = await CourseService.getCourseById(idCourse);
       res.status(200).json(data);
     } catch (error: any) {
-      throw new Error(error.message);
+      throw new CustomError({
+        message: 'Error al buscar curso por id',
+        status: 500,
+        name: 'courseById',
+      });
     }
   }
 
   static async deleteCourse(req: Request, res: Response) {
     try {
-      const idCourse = Number(req.params.id);
-      await CourseService.deleteCourse(idCourse);
-      res.status(204).json();
+      const user: rol = req.body.userlogued.rol;
+      if (user !== rol.admin) {
+        res.status(403).json({ msg: 'no tiene permiso de admnistrador' });
+      } else {
+        const idCourse = Number(req.params.id);
+        await CourseService.deleteCourse(idCourse);
+        res.status(204).json();
+      }
     } catch (error: any) {
-      throw new Error(error.message);
+      console.log(error);
+      throw new CustomError({
+        message: 'Error al buscar borrar curso',
+        status: 500,
+        name: 'deleteCourse',
+      });
     }
   }
 
@@ -39,7 +47,11 @@ export class CourseController {
       const data = await CourseService.updateCourse(idCourse, course);
       res.status(200).json(data);
     } catch (error: any) {
-      throw new Error(error.message);
+      throw new CustomError({
+        message: 'Error al actualizar un curso',
+        status: 500,
+        name: 'updateCourse',
+      });
     }
   }
 
@@ -49,26 +61,24 @@ export class CourseController {
       const data = await CourseService.createCourse(course);
       res.status(200).json(data);
     } catch (error: any) {
-      throw new Error(error.message);
+      throw new CustomError({
+        message: 'Error al crear un curso',
+        status: 500,
+        name: 'createCourse',
+      });
     }
   }
 
-  static async getCourseByTheme(req: Request, res: Response) {
+  static async getCourses(req: Request, res: Response) {
     try {
-      const theme = (req.params.name as string) || '';
-      const data = await CourseService.getCourseByTheme(theme);
+      const data = await CourseService.getCourses(req);
       res.status(200).json(data);
     } catch (error: any) {
-      throw new Error(error.message);
-    }
-  }
-
-  static async getCoursesByThemeAndPrice(req: Request, res: Response) {
-    try {
-      const data = await CourseService.getCoursesByThemeAndPrice(req);
-      res.status(200).json(data);
-    } catch (error: any) {
-      throw new Error(error.message);
+      throw new CustomError({
+        message: 'Error al consultar cursos',
+        status: 500,
+        name: 'getCourses',
+      });
     }
   }
 }
